Start using feed hashing
As it stands on every poll cycle every feed is being processed the
whole way down the chain. 304's mitigate this but not everyone uses
304's. So, we are going to do something about that on our side.

Everytime we fetch a feed we are going to hash the content. If the
content hash is the same as the last fetch we skip all the feed
processing. If the hash's differ we update the database, and continue
on looking for new items.